Move test case of auth and push to separate file
